% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cross_validation_learning_algorithms}
\alias{cross_validation_learning_algorithms}
\title{Cross validation for learning bayesian network}
\usage{
cross_validation_learning_algorithms(
  data,
  algorithms,
  wl,
  bl,
  target,
  k_fold = 5,
  runs = 10,
  loss = "mse",
  n_cluster = NULL,
  debug = NULL
)
}
\arguments{
\item{data}{Processed data frame for learning bayesian network}

\item{algorithms}{Algorithms for cross validation}

\item{wl}{Network white list}

\item{bl}{Network black list}

\item{target}{A character string, the label of target node for prediction}

\item{k_fold}{The data are split in k subsets of equal size}

\item{runs}{A positive integer number, the number of times k-fold or hold-out cross-validation will be run.}

\item{loss}{A character string, the label of a loss function, detail see https://cran.r-project.org/web/packages/bnlearn/bnlearn.pdf}

\item{n_cluster}{an optional cluster object from package parallel.}

\item{debug}{Debug mode}
}
\value{
Cross validation result for each algorithm
}
\description{
Cross validation for learning bayesian network
}
