% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_continuous_structure_filter}
\alias{get_continuous_structure_filter}
\title{Filter blacklist and whitelist before training bayesian network}
\usage{
get_continuous_structure_filter(
  continuous_data,
  desire_layers,
  quantile_number,
  continuous_variables,
  discrete_variables,
  known_structure,
  corr_threshold,
  is_blacklist_internal,
  is_variable_only,
  is_blacklist_other,
  custom_blacklist,
  custom_whitelist
)
}
\arguments{
\item{continuous_data}{A data frame which rows are financial reports and columns is KPI pre-processed}

\item{desire_layers}{Number of layers for training bayesian network, less than or equal current layers}

\item{quantile_number}{For type is "discrete", number of quantile for each continuous variable}

\item{continuous_variables}{Column names of continuous dynamic KPI}

\item{discrete_variables}{Column names of discrete dynamic KPI}

\item{known_structure}{Relation between KPI specific by user}

\item{corr_threshold}{Threshold for filter arcs with low correlation}

\item{is_blacklist_internal}{Is blacklist all internal arcs in n-i layers (n is number layers, i = 1..n-1)}

\item{is_variable_only}{Is known_structure is KPI only, without time notation}

\item{is_blacklist_other}{Is blacklist all except whitelist}

\item{custom_blacklist}{Blacklist specific by user}

\item{custom_whitelist}{Whitelist sepcific by user}
}
\value{
An object with blacklist and whitelist of network
}
\description{
Filter blacklist and whitelist before training bayesian network
}
\examples{
data("preprocessed")
data("data_small")
continuous_data <- preprocessed$data
desire_layers <- preprocessed$desire_layers
quantile_number <- preprocessed$quantile_number
continuous_dynamic_variables <- preprocessed$continuous_dynamic_variables
continuous_static_variables <- preprocessed$continuous_static_variables
discrete_static_variables <- preprocessed$discrete_static_variables
known_structure <- NULL
corr_threshold <- 0.4
is_blacklist_internal <- TRUE
is_variable_only <- TRUE
is_blacklist_other <- FALSE
custom_blacklist <- NULL
custom_whitelist <- NULL
bl_wl <- get_continuous_structure_filter(continuous_data, desire_layers, quantile_number, continuous_dynamic_variables,
                                                continuous_static_variables, discrete_static_variables,
                                                known_structure, corr_threshold, is_blacklist_internal,
                                                is_variable_only, is_blacklist_other,
                                                custom_blacklist, custom_whitelist)
}
