% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{impute_missing_data}
\alias{impute_missing_data}
\title{Impute missing data in data frame}
\usage{
impute_missing_data(
  fitted,
  data,
  continuous_variables,
  discrete_variables,
  time_column = NULL,
  normalize_type = NULL,
  normalizers = NULL,
  debug = FALSE
)
}
\arguments{
\item{fitted}{A fitted model}

\item{data}{A data frame, contains missing data, each row is a time value and observations, structure is the same as learning data before process}

\item{continuous_variables}{Column names of continuous variables}

\item{discrete_variables}{Column names of discrete variables}

\item{time_column}{Column name of "data", which values is time stamp, default is NULL}

\item{normalize_type}{Normalization type for continuous variables, "mean_normalization", "min_max" or "standardisation", default is NULL}

\item{normalizers}{Normalize parameters, default is NULL}
}
\value{
A full filled data frame
}
\description{
Impute missing data in data frame
}
\examples{
library(wrmbn)
data("data")
data("preprocessed")
data("trained_models")
head(data)
continuous_variables <- preprocessed$continuous_variables
discrete_variables <- preprocessed$discrete_variables
desire_layers <- preprocessed$desire_layers
time_column <- "date"
normalize_type <- preprocessed$normalize_tye
normalizers <- preprocessed$normalizers
fitted <- trained_models$hc$fitted

imputed_data <- impute_missing_data(fitted, data, continuous_variables, discrete_variables,
                                    time_column, normalize_type, normalizers, debug = FALSE)
head(imputed_data)
}
