% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict_data}
\alias{predict_data}
\title{Predict data}
\usage{
predict_data(
  fitted,
  predictor_data,
  predicted_nodes,
  number_layers,
  continuous_variables,
  discrete_variables,
  time_column = NULL,
  time_format = "\%m/\%d/\%y",
  normalizers = NULL,
  normalize_type = NULL,
  method = "lw",
  debug = FALSE
)
}
\arguments{
\item{fitted}{A fitted model}

\item{predictor_data}{A data frame, each row is evidence}

\item{predicted_nodes}{List of variables need to predict}

\item{number_layers}{Number layers of fitted model}

\item{continuous_variables}{Column names of continuous variables}

\item{discrete_variables}{Column names of discrete variables}

\item{time_column}{Column name of "data", which values is time stamp, default is NULL}

\item{time_format}{Time format of time column, default is %m/%d/%y}

\item{normalizers}{Normalize parameters, default is NULL}

\item{normalize_type}{Normalization type for continuous variables, "mean_normalization", "min_max" or "standardisation", default is NULL}
}
\value{
A data frame with evidence and predicted variables
}
\description{
Predict data
}
\examples{
library(wrmbn)
data("data")
data("preprocessed")
data("trained_models")
predictor_data <- data[2019:2020, c("date", "HND", "HCT")]
print(predictor_data)
normalizers <- preprocessed$normalizers
normalize_type <- preprocessed$normalize_tye
fitted <- trained_models$hc$fitted
continuous_variables <- c("HND", "HCT")
discrete_variables <- c()
predicted_nodes <- c("MBT")
number_layers <- 3

predicted_data <- predict_data(fitted, predictor_data, predicted_nodes, number_layers,
                               continuous_variables, discrete_variables,
                               time_column = "date", "\%m/\%d/\%y", normalizers, normalize_type,
                               method = "lw", TRUE)

print(predicted_data)
actual_data <- data[2019:2021, c("date", "HND", "HCT", "MBT")]
print(actual_data)
}
