% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{preprocess_test_data}
\alias{preprocess_test_data}
\title{Preprocess time series data prepare for test learned bayesian network}
\usage{
preprocess_test_data(
  data,
  continuous_variables,
  discrete_variables,
  desire_layers,
  time_column = NULL,
  normalize_type = NULL,
  normalizers = NULL
)
}
\arguments{
\item{data}{A data frame, each row is a time value and observations}

\item{desire_layers}{Number layers of bayesian network, at least is 2}

\item{time_column}{Column name of "data", which values is time stamp, default is NULL}

\item{normalize_type}{Normalization type for continuous variables, "mean_normalization", "min_max" or "standardisation", default is NULL}

\item{normalizers}{Normalize parameters, default is NULL}

\item{continuous_variable_names}{Column names of continuous variables}

\item{discrete_variable_names}{Column names of discrete variables}
}
\value{
An list object, each element is processed data frame
}
\description{
Preprocess time series data prepare for test learned bayesian network
}
\examples{
library(wrmbn)
data("preprocessed")
head(data)
continuous_variables <- preprocessed$continuous_variables
discrete_variables <- preprocessed$discrete_variables
desire_layers <- preprocessed$desire_layers
time_column <- "date"
normalize_type <- preprocessed$normalize_tye
normalizers <- preprocessed$normalizers
test_data <- preprocess_test_data(data, continuous_variables, discrete_variables, desire_layers,
                                  time_column, normalize_type, normalizers)
}
