% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{preprocess_timeseries_data}
\alias{preprocess_timeseries_data}
\title{Preprocess time series data prepare for learning bayesian network}
\usage{
preprocess_timeseries_data(
  data,
  type,
  time_column,
  continuous_variable_names,
  discrete_variable_names,
  desire_layers,
  normalize_type = NULL,
  quantile_number = -1,
  na_omit = TRUE
)
}
\arguments{
\item{data}{A data frame, each row is a time value and observations}

\item{type}{Bayesian network type, "discrete" or "continuous"}

\item{time_column}{Column name of "data", which values is time stamp}

\item{continuous_variable_names}{Column names of continuous variables}

\item{discrete_variable_names}{Column names of discrete variables}

\item{desire_layers}{Number layers of bayesian network, at least is 2}

\item{normalize_type}{Normalization type for continuous variables, "mean_normalization", "min_max" or "standardisation"}

\item{quantile_number}{Number of quantile level for type "discrete"}

\item{na_omit}{If true, NA/NaN values will be omit after preprocess}
}
\value{
An object with pre-processsed data and preprocess parameters
}
\description{
Preprocess time series data prepare for learning bayesian network
}
\examples{
library(wrmbn)

data("data")
data("structure")

uncorelate_node <- c("MTL", "QMX", "HCL", "CMB", "CTL", "CDX", "CBT")
for(node in uncorelate_node) {
 data <- data[, -which(colnames(data) == node)]
}

for(node in uncorelate_node) {
  if(length(which(structure$from == node)) > 0) {
    structure <- structure[-which(structure$from == node), ]
  } else if(length(which(structure$to == node)) > 0) {
    structure <- structure[-which(structure$to == node), ]
  }
}
type <- "continuous"
time_column <- "date"
continuous_variable_names <- setdiff(colnames(data), time_column)
discrete_variable_names <- c()
desire_layers <- 3
normalize_type <- "mean_normalization"
preprocessed <- preprocess_timeseries_data(data, type, time_column,
                                           continuous_variable_names, discrete_variable_names, desire_layers,
                                           normalize_type, quantile_number = -1, na_omit = TRUE)
}
